apps/api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
requests==2.32.4
redis==6.4.0
beautifulsoup4==4.12.3
python-dotenv==1.1.1
